+++ cypress/integration/blocks.js
@@ -476,30 +476,79 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.get('#page-document pre').should('have.text', 'This is HTML');
     });
 
-    // it('Add table block', () => {
-    // TODO: Figure out why there is an erro when add this block in cypress
-    // const block = 'table';
-    // const expected = 'This is the html';
-    // Edit
-    // cy.get('.block.text [contenteditable]').click();
-    // cy.get('button.block-add-button').click();
-    // cy.get('button.show-hidden-blocks').click();
-    // cy.get(`button.add-${block}-block`).click();
-    // cy.get(`.block.${block} .npm__react-simple-code-editor__textarea`).type(
-    //   `<h3>${expected}</h3>`,
-    // );
-    // cy.get(`.block.${block} [aria-label="Preview"]`).click();
-    // cy.get(`.block.${block} h3`).contains(expected);
-    // // Save
-    // cy.get('#toolbar-save').click();
-    // // View
-    // if (Cypress.env('API') === 'plone') {
-    //   cy.get('#page-document h3').should('have.text', expected);
-    // } else {
-    //   // guillotina
-    //   cy.contains(expected);
-    // }
-    // });
+    it('Add table block', () => {
+      // Edit
+      cy.get('.block.text [contenteditable]').click();
+      cy.get('button.block-add-button').click();
+      cy.get('.blocks-chooser .title')
+        .contains('Common')
+        .click();
+      cy.get('.ui.buttons .button.table').click();
+      cy.get('.celled.fixed.table tr th:first-child()')
+        .click()
+        .type('column 1 / row 1');
+      cy.get('.celled.fixed.table tr th:nth-child(2)')
+        .click()
+        .type('column 2 / row 1');
+      cy.get('.celled.fixed.table tr:nth-child(2) td:first-child()')
+        .click()
+        .type('column 1 / row 2');
+      cy.get('.celled.fixed.table tr:nth-child(2) td:nth-child(2)')
+        .click()
+        .type('column 2 / row 2');
+      cy.get('button[title="Insert col after"]').click();
+      cy.get('button[title="Insert row after"]').click();
+      cy.get('button[title="Insert row before"]').click();
+      cy.get('button[title="Insert col before"]').click();
+
+      // Save
+      cy.get('#toolbar-save').click();
+
+      // View
+      cy.get('.celled.fixed.table tr th:first-child()').contains(
+        'column 1 / row 1',
+      );
+      cy.get('.celled.fixed.table tr th:nth-child(3)').contains(
+        'column 2 / row 1',
+      );
+      cy.get('.celled.fixed.table tr:nth-child(3) td:first-child()').contains(
+        'column 1 / row 2',
+      );
+      cy.get('.celled.fixed.table tr:nth-child(3) td:nth-child(3)').contains(
+        'column 2 / row 2',
+      );
+
+      // Edit
+      cy.visit('/my-page/edit');
+      cy.get('.celled.fixed.table tr:first-child() th:nth-child(2)').click();
+
+      // without the second click the test fails. so this makes the test green.
+      cy.get('.celled.fixed.table tr:first-child() th:nth-child(2)').click();
+
+      cy.get('button[title="Delete col"]').click();
+      cy.get('.celled.fixed.table tr:first-child() th:nth-child(3)').click();
+      cy.get('button[title="Delete col"]').click();
+      cy.get('.celled.fixed.table tr:nth-child(2) td:first-child()').click();
+      cy.get('button[title="Delete row"]').click();
+      cy.get('.celled.fixed.table tr:nth-child(3) td:first-child()').click();
+      cy.get('button[title="Delete row"]').click();
+
+      // Save
+      cy.get('#toolbar-save').click();
+      cy.url().should('eq', Cypress.config().baseUrl + '/my-page');
+
+      //View
+      cy.get('.celled.fixed.table tr th:first-child()').contains(
+        'column 1 / row 1',
+      );
+      cy.get('th:nth-child(2)> p ').should('have.text', 'column 2 / row 1');
+      cy.get('.celled.fixed.table tr:nth-child(2) td:first-child()').contains(
+        'column 1 / row 2',
+      );
+      cy.get('.celled.fixed.table tr:nth-child(2) td:nth-child(2)').contains(
+        'column 2 / row 2',
+      );
+    });
 
     it('Add Table of Contents block', () => {
       // given a text block with a H2 headline
@@ -527,7 +576,7 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.waitForResourceToLoad('@breadcrumbs');
       cy.waitForResourceToLoad('@actions');
       cy.waitForResourceToLoad('@types');
-      cy.waitForResourceToLoad('my-page?fullobjects');
+      cy.waitForResourceToLoad('?fullobjects');
 
       // then the ToC block should contain the H2 headline
       cy.get('.block.table-of-contents .ui.list a').contains(